+++ baybe/utils/interval.py
@@ -1,6 +1,7 @@
 # pylint: disable=missing-function-docstring, missing-module-docstring
 
 import sys
+from collections.abc import Iterable
 from functools import singledispatchmethod
 from typing import Union
 
@@ -70,7 +71,7 @@ class Interval:
 
     @create.register
     @classmethod
-    def _(cls, bounds: tuple):
+    def _(cls, bounds: Iterable):
         return Interval(*bounds)
 
     def to_tuple(self):